+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed CPU target specifications because their feature mappings could not be kept consistent as rustc's feature database changed.
 ### Fixed
 - Disabling the `std` feature now makes the crate `no_std`.
-- Fixed handling of `'static` and `'_` lifetimes in multiversioned function signatures.
+- Fixed handling of `'static`, `'_`, and bounded lifetimes in multiversioned function signatures.
 - Fixed compilation of multiversioned `unsafe fn` bodies containing implicit unsafe operations.
 
 ## [0.8.0] - 2024-12-07